// src/views/swap/Widget2/lib/uniswap/src/data/graphql/uniswap-data-api/__generated__/types-and-hooks.ts
export type Maybe<T> = T | undefined
// ...
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /**
   * The `AWSJSON` scalar type provided by AWS AppSync, represents a JSON string that
   * complies with [RFC 8259](https://tools.ietf.org/html/rfc8259).  Maps like
   * "**{\\"upvotes\\": 10}**", lists like "**[1,2,3]**", and scalar values like
   * "**\\"AWSJSON example string\\"**", "**1**", and "**true**" are accepted as
   * valid JSON and will automatically be parsed and loaded in the resolver mapping
   * templates as Maps, Lists, or Scalar values rather than as the literal input
   * strings.  Invalid JSON strings like "**{a: 1}**", "**{'a': 1}**" and "**Unquoted
   * string**" will throw GraphQL validation errors.
   */
  AWSJSON: any
}
// ...
export type Amount = IAmount & {
  __typename?: 'Amount'
  currency?: Maybe<Currency>
  id: Scalars['ID']
  value: Scalars['Float']
}
// ...
export enum Chain {
  Arbitrum = 'ARBITRUM',
  Avalanche = 'AVALANCHE',
  Base = 'BASE',
  Blast = 'BLAST',
  Bnb = 'BNB',
  Celo = 'CELO',
  Ethereum = 'ETHEREUM',
  EthereumGoerli = 'ETHEREUM_GOERLI',
  EthereumSepolia = 'ETHEREUM_SEPOLIA',
  Optimism = 'OPTIMISM',
  Polygon = 'POLYGON',
  UnknownChain = 'UNKNOWN_CHAIN',
  BounceBit = 'BOUNCEBIT'
}

export enum CollectionSortableField {
  Volume = 'VOLUME'
}

export type ContractInput = {
  address?: InputMaybe<Scalars['String']>
  chain: Chain
}

export enum Currency {
  Aud = 'AUD',
  Brl = 'BRL',
  Cad = 'CAD',
  Cny = 'CNY',
  Eth = 'ETH',
  Eur = 'EUR',
  Gbp = 'GBP',
  Hkd = 'HKD',
  Idr = 'IDR',
  Inr = 'INR',
  Jpy = 'JPY',
  Matic = 'MATIC',
  Ngn = 'NGN',
  Pkr = 'PKR',
  Rub = 'RUB',
  Sgd = 'SGD',
  Thb = 'THB',
  Try = 'TRY',
  Uah = 'UAH',
  Usd = 'USD',
  Vnd = 'VND'
}

export type CurrencyAmountInput = {
  currency: Currency
  value: Scalars['Float']
}

export type DescriptionTranslations = {
  __typename?: 'DescriptionTranslations'
  descriptionEnUs?: Maybe<Scalars['String']>
  descriptionEs419?: Maybe<Scalars['String']>
  descriptionEsEs?: Maybe<Scalars['String']>
  descriptionEsUs?: Maybe<Scalars['String']>
  descriptionFrFr?: Maybe<Scalars['String']>
  descriptionHiIn?: Maybe<Scalars['String']>
  descriptionIdId?: Maybe<Scalars['String']>
  descriptionJaJp?: Maybe<Scalars['String']>
  descriptionMsMy?: Maybe<Scalars['String']>
  descriptionNlNl?: Maybe<Scalars['String']>
  descriptionPtPt?: Maybe<Scalars['String']>
  descriptionRuRu?: Maybe<Scalars['String']>
  descriptionThTh?: Maybe<Scalars['String']>
  descriptionTrTr?: Maybe<Scalars['String']>
  descriptionUkUa?: Maybe<Scalars['String']>
  descriptionUrPk?: Maybe<Scalars['String']>
  descriptionViVn?: Maybe<Scalars['String']>
  descriptionZhHans?: Maybe<Scalars['String']>
  descriptionZhHant?: Maybe<Scalars['String']>
  id: Scalars['ID']
}

export type Dimensions = {
  __typename?: 'Dimensions'
  height?: Maybe<Scalars['Float']>
  id: Scalars['ID']
  width?: Maybe<Scalars['Float']>
}

export type DimensionsInput = {
  height?: InputMaybe<Scalars['Float']>
  width?: InputMaybe<Scalars['Float']>
}

export type FeeData = {
  __typename?: 'FeeData'
  buyFeeBps?: Maybe<Scalars['String']>
  sellFeeBps?: Maybe<Scalars['String']>
}

export enum HighLow {
  High = 'HIGH',
  Low = 'LOW'
}

/**   FIVE_MINUTE is only supported for TokenMarket.pricePercentChange */
export enum HistoryDuration {
  Day = 'DAY',
  FiveMinute = 'FIVE_MINUTE',
  Hour = 'HOUR',
  Max = 'MAX',
  Month = 'MONTH',
  Week = 'WEEK',
  Year = 'YEAR'
}

/**   Interfaces (alphabetized): */
export type IAmount = {
  currency?: Maybe<Currency>
  value: Scalars['Float']
}

export type IContract = {
  address?: Maybe<Scalars['String']>
  chain: Chain
}
// ...
export type Image = {
  __typename?: 'Image'
  dimensions?: Maybe<Dimensions>
  id: Scalars['ID']
  url: Scalars['String']
}
// ...
export type PoolTransaction = {
  __typename?: 'PoolTransaction'
  account: Scalars['String']
  chain: Chain
  hash: Scalars['String']
  id: Scalars['ID']
  protocolVersion: ProtocolVersion
  timestamp: Scalars['Int']
  token0: Token
  token0Quantity: Scalars['String']
  token1: Token
  token1Quantity: Scalars['String']
  type: PoolTransactionType
  usdValue: Amount
}
// ...
export enum SafetyLevel {
  Blocked = 'BLOCKED',
  MediumWarning = 'MEDIUM_WARNING',
  StrongWarning = 'STRONG_WARNING',
  Verified = 'VERIFIED'
}
// ...
export type TimestampedAmount = IAmount & {
  __typename?: 'TimestampedAmount'
  currency?: Maybe<Currency>
  id: Scalars['ID']
  timestamp: Scalars['Int']
  value: Scalars['Float']
}

export type TimestampedOhlc = {
  __typename?: 'TimestampedOhlc'
  close: Amount
  high: Amount
  id: Scalars['ID']
  low: Amount
  open: Amount
  timestamp: Scalars['Int']
}

export type TimestampedPoolPrice = {
  __typename?: 'TimestampedPoolPrice'
  id: Scalars['ID']
  timestamp: Scalars['Int']
  token0Price: Scalars['Float']
  token1Price: Scalars['Float']
}

export type Token = IContract & {
  __typename?: 'Token'
  address?: Maybe<Scalars['String']>
  chain: Chain
  decimals?: Maybe<Scalars['Int']>
  feeData?: Maybe<FeeData>
  id: Scalars['ID']
  market?: Maybe<TokenMarket>
  name?: Maybe<Scalars['String']>
  project?: Maybe<TokenProject>
  standard?: Maybe<TokenStandard>
  symbol?: Maybe<Scalars['String']>
  v2Transactions?: Maybe<Array<Maybe<PoolTransaction>>>
  v3Transactions?: Maybe<Array<Maybe<PoolTransaction>>>
}
// ...
export type TokenBalance = {
  __typename?: 'TokenBalance'
  blockNumber?: Maybe<Scalars['Int']>
  blockTimestamp?: Maybe<Scalars['Int']>
  denominatedValue?: Maybe<Amount>
  id: Scalars['ID']
  isHidden?: Maybe<Scalars['Boolean']>
  ownerAddress: Scalars['String']
  quantity?: Maybe<Scalars['Float']>
  token?: Maybe<Token>
  tokenProjectMarket?: Maybe<TokenProjectMarket>
}

export type TokenInput = {
  address: Scalars['String']
  chainId: Scalars['Int']
  decimals: Scalars['Int']
  isNative: Scalars['Boolean']
}

export type TokenMarket = {
  __typename?: 'TokenMarket'
  fullyDilutedValuation?: Maybe<Amount>
  historicalTvl?: Maybe<Array<Maybe<TimestampedAmount>>>
  historicalVolume?: Maybe<Array<Maybe<TimestampedAmount>>>
  id: Scalars['ID']
  ohlc?: Maybe<Array<Maybe<TimestampedOhlc>>>
  price?: Maybe<Amount>
  priceHighLow?: Maybe<Amount>
  priceHistory?: Maybe<Array<Maybe<TimestampedAmount>>>
  pricePercentChange?: Maybe<Amount>
  priceSource: PriceSource
  token: Token
  totalValueLocked?: Maybe<Amount>
  /**   this volume is cumulative volume over the specified duration */
  volume?: Maybe<Amount>
}
// ...
export type TokenProject = {
  __typename?: 'TokenProject'
  description?: Maybe<Scalars['String']>
  descriptionTranslations?: Maybe<DescriptionTranslations>
  homepageUrl?: Maybe<Scalars['String']>
  id: Scalars['ID']
  isSpam?: Maybe<Scalars['Boolean']>
  logo?: Maybe<Image>
  /** @deprecated use logo */
  logoUrl?: Maybe<Scalars['String']>
  markets?: Maybe<Array<Maybe<TokenProjectMarket>>>
  name?: Maybe<Scalars['String']>
  safetyLevel?: Maybe<SafetyLevel>
  /** @deprecated use logo */
  smallLogo?: Maybe<Image>
  spamCode?: Maybe<Scalars['Int']>
  tokens: Array<Token>
  twitterName?: Maybe<Scalars['String']>
}

export type TokenProjectMarketsArgs = {
  currencies: Array<Currency>
}

export type TokenProjectMarket = {
  __typename?: 'TokenProjectMarket'
  currency: Currency
  fullyDilutedValuation?: Maybe<Amount>
  id: Scalars['ID']
  marketCap?: Maybe<Amount>
  price?: Maybe<Amount>
  priceHigh52w?: Maybe<Amount>
  priceHighLow?: Maybe<Amount>
  priceHistory?: Maybe<Array<Maybe<TimestampedAmount>>>
  priceLow52w?: Maybe<Amount>
  pricePercentChange?: Maybe<Amount>
  pricePercentChange24h?: Maybe<Amount>
  tokenProject: TokenProject
}
// ...
export enum TokenStandard {
  Erc20 = 'ERC20',
  Native = 'NATIVE'
}
// ...
export type PortfolioTokenBalancePartsFragment = {
  __typename?: 'TokenBalance'
  id: string
  quantity?: number | undefined
  denominatedValue?: { __typename?: 'Amount'; id: string; currency?: Currency | undefined; value: number } | undefined
  token?:
    | {
        __typename?: 'Token'
        id: string
        chain: Chain
        address?: string | undefined
        name?: string | undefined
        symbol?: string | undefined
        standard?: TokenStandard | undefined
        decimals?: number | undefined
      }
    | undefined
  tokenProjectMarket?:
    | {
        __typename?: 'TokenProjectMarket'
        id: string
        pricePercentChange?: { __typename?: 'Amount'; id: string; value: number } | undefined
        tokenProject: {
          __typename?: 'TokenProject'
          id: string
          logoUrl?: string | undefined
          isSpam?: boolean | undefined
        }
      }
    | undefined
}